# docs/swagger.yaml
basePath: /api
definitions:
...
  v1.ValidateEmailJsonForm:
    properties:
      validatorCode:
        type: string
      validatorKey:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: All APIs of Gol-C Project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gol-C API
  version: "1.0"
paths:
  /v1/auth/email-validations/any:validate:
    put:
      description: Validate a user's email
      parameters:
      - description: Email validation from
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/v1.ValidateEmailJsonForm'
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: ValidateEmail
      tags:
      - auth
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"